parser/tests/tree_construction/basic.dat

#data
<a><div><p></a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,15): adoption-agency-1.3
(1,15): adoption-agency-1.3
(1,15): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|     <div>
|       <a>
|       <p>
|         <a>
